v0.0.7

+++ src/backend/java/org/example/Angular/servies/CategoryService.java
@@ -4,7 +4,9 @@ import org.example.Angular.entities.Category;
 import org.example.Angular.repositories.CategoryRepository;
 import org.springframework.stereotype.Service;
 
+import javax.annotation.PostConstruct;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CategoryService {
@@ -28,6 +30,19 @@ public class CategoryService {
     }
 
     public void deleteCategory(Category category){
+        Optional<Category> defaultCategory = categoryRepository.findById(1);
+        if(!defaultCategory.isPresent())
+            return;
+
+        category.getSubCategories().forEach(x -> x.setCategory(defaultCategory.get()));
         categoryRepository.delete(category);
+    }
+
+    @PostConstruct
+    public void onInit(){
+        if(categoryRepository.findByName("Default").isPresent())
+            return;
+
+        categoryRepository.save(new Category("Default"));
     }
 }
